+++ migrations/8_oauth.up.sql
@@ -3,6 +3,7 @@ CREATE TABLE IF NOT EXISTS `oauth_tokens` (
     `service` VARCHAR(64) NOT NULL,
     `access_token` VARCHAR(255) NULL DEFAULT '',
     `refresh_token` VARCHAR(255) NULL DEFAULT '',
+    `url` VARCHAR(255) NULL DEFAULT '',
     `expiry` DATETIME NOT NULL,
     `username` VARCHAR(100) NULL DEFAULT '',
     `last_synced` DATETIME NOT NULL DEFAULT NOW(),