+++ web/src/Components/ScrobbleTable.js
@@ -3,15 +3,14 @@ import { Link } from 'react-router-dom';
 
 const ScrobbleTable = (props) => {
     return (
-        <div>
-        <table width={900} border={1} cellPadding={5}>
+        <div style={{width: `100%`}}>
+        <table style={{width: `100%`}} border={1} cellPadding={5}>
            <thead>
               <tr>
                 <td>Timestamp</td>
                 <td>Track</td>
                 <td>Artist</td>
                 <td>Album</td>
-                <td>Source</td>
               </tr>
            </thead>
            <tbody>
@@ -30,7 +29,6 @@ const ScrobbleTable = (props) => {
                         </td>
                         <td>{element.artist}</td>
                         <td>{element.album}</td>
-                        <td>{element.source}</td>
                      </tr>;
                   })
               }